# src/main/java/com/nazarov/projects/blog/models/ErrorResponse.java
package com.nazarov.projects.blog.models;

import java.util.List;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder(toBuilder = true)
public class ErrorResponse {

  private List<String> message;
}