# src/main/java/com/nazarov/projects/blog/models/dtos/TagDto.java
package com.nazarov.projects.blog.models.dtos;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Builder(toBuilder = true)
public class TagDto {

  private Long id;
  private String name;

}